+++ src/app/index.tsx
@@ -2,7 +2,7 @@ import MainPage from '@/pages/MainPage/MainPage';
 import { HeroUIProvider } from '@heroui/system';
 import { FC } from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import MoviesResponseContextProvider from './providers/MoviesResponseContextProvider/context';
+import MoviesResponseContextProvider from '../entities/providers/MoviesResponseContextProvider/context';
 
 export const App: FC = () => (
   <BrowserRouter>
