+++ src/features/SearchMoviesInput/SearchMoviesInput.tsx
@@ -3,22 +3,22 @@ import { Input } from '@heroui/input';
 import { Button } from '@heroui/button';
 import { Form } from '@heroui/form';
 import { MagnifyingGlassIcon } from '@heroicons/react/24/solid';
-import { getMovies } from '../api/getMovies';
-import { MoviesResponseContext } from '@/app/providers/MoviesResponseContextProvider/context';
+import { MoviesResponseContext } from '@/entities/providers/MoviesResponseContextProvider/context';
 
-const SearchMoviesInput: FC = () => {
-  const { moviesList, addMovies } = useContext(MoviesResponseContext);
+export const SearchMoviesInput: FC = () => {
+  const { moviesList, addMovies, getMovies } = useContext(
+    MoviesResponseContext,
+  );
   const onSearcheMovie = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     const form = e.currentTarget;
     const formData = new FormData(e.currentTarget);
     const title = formData.get('search') as string;
     const moviesData = await getMovies(title);
+    console.log(moviesData);
     addMovies(moviesData);
     console.log(moviesList);
-    if (form && typeof form.reset === 'function') {
-      form.reset();
-    }
+    form?.reset();
   };
 
   return (
@@ -47,5 +47,3 @@ const SearchMoviesInput: FC = () => {
     </Form>
   );
 };
-
-export default SearchMoviesInput;
