+++ src/widgets/Navbar/ui/Navbar.tsx
@@ -8,7 +8,7 @@ import {
 import { FilmIcon } from '@heroicons/react/24/outline';
 import { FC } from 'react';
 import { ThemeSwitch } from '@/features/ThemeSwitch/ThemeSwitch';
-import SearchInput from '@/features/SearchMoviesInput/ui/SearchMoviesInput';
+import { SearchMoviesInput } from '@/features/SearchMoviesInput/SearchMoviesInput';
 
 export const Navbar: FC = () => {
   return (
@@ -25,7 +25,7 @@ export const Navbar: FC = () => {
           </Link>
         </NavbarBrand>
         <NavbarItem className="flex items-center gap-4 flex-grow">
-          <SearchInput />
+          <SearchMoviesInput />
           <ThemeSwitch />
         </NavbarItem>
       </NavbarContent>
